requirements.txt
# requirements.txt
pandas==2.1.2
minio==7.2.14
apache-airflow-providers-amazon==8.29.0
apache-airflow-providers-trino==6.0.0
apache-airflow-providers-common-compat==1.2.2
apache-airflow-providers-apache-spark==4.11.3